+++ src/main/java/com/goggles/mentoring_service/application/command/MentoringCommand.java
@@ -1,7 +1,7 @@
 package com.goggles.mentoring_service.application.command;
 
 import com.goggles.mentoring_service.domain.category.MentoringCategory;
-import com.goggles.mentoring_service.domain.common.UserType;
+import com.goggles.mentoring_service.domain._common.UserType;
 import com.goggles.mentoring_service.domain.mentoring.*;
 
 import java.time.LocalDate;
