+++ src/main/java/com/goggles/mentoring_service/domain/_common/MentoringDomainException.java
@@ -1,4 +1,4 @@
-package com.goggles.mentoring_service.domain.common;
+package com.goggles.mentoring_service.domain._common;
 
 public abstract class MentoringDomainException extends RuntimeException {
