+++ src/main/java/com/goggles/mentoring_service/domain/_common/UserType.java
@@ -1,4 +1,4 @@
-package com.goggles.mentoring_service.domain.common;
+package com.goggles.mentoring_service.domain._common;
 
 public enum UserType {
 	STUDENT, INSTRUCTOR, MASTER
