+++ src/main/java/com/goggles/mentoring_service/domain/booking/Mentee.java
@@ -1,7 +1,7 @@
 package com.goggles.mentoring_service.domain.booking;
 
 import com.goggles.mentoring_service.domain.booking.exception.InvalidMenteeUserTypeException;
-import com.goggles.mentoring_service.domain.common.UserType;
+import com.goggles.mentoring_service.domain._common.UserType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
