+++ src/main/java/com/goggles/mentoring_service/domain/booking/MentoringBooking.java
@@ -4,7 +4,7 @@ import com.goggles.common.domain.BaseAudit;
 import com.goggles.mentoring_service.domain.booking.exception.CancellationDeadlineExceededException;
 import com.goggles.mentoring_service.domain.booking.exception.CancellationReasonRequiredException;
 import com.goggles.mentoring_service.domain.booking.exception.UnauthorizedBookingAccessException;
-import com.goggles.mentoring_service.domain.common.UserType;
+import com.goggles.mentoring_service.domain._common.UserType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
