+++ src/main/java/com/goggles/mentoring_service/domain/booking/exception/CancellationReasonRequiredException.java
@@ -1,6 +1,6 @@
 package com.goggles.mentoring_service.domain.booking.exception;
 
-import com.goggles.mentoring_service.domain.common.MentoringValidationException;
+import com.goggles.mentoring_service.domain._common.MentoringValidationException;
 
 public class CancellationReasonRequiredException extends MentoringValidationException {
