+++ src/main/java/com/goggles/mentoring_service/domain/booking/exception/InvalidMenteeUserTypeException.java
@@ -1,7 +1,7 @@
 package com.goggles.mentoring_service.domain.booking.exception;
 
-import com.goggles.mentoring_service.domain.common.MentoringValidationException;
-import com.goggles.mentoring_service.domain.common.UserType;
+import com.goggles.mentoring_service.domain._common.MentoringValidationException;
+import com.goggles.mentoring_service.domain._common.UserType;
 
 import java.util.UUID;
