+++ src/main/java/com/goggles/mentoring_service/domain/category/exception/CategoryAdminForbiddenException.java
@@ -1,7 +1,7 @@
 package com.goggles.mentoring_service.domain.category.exception;
 
 import com.goggles.common.exception.ForbiddenException;
-import com.goggles.mentoring_service.domain.common.UserType;
+import com.goggles.mentoring_service.domain._common.UserType;
 
 import java.util.UUID;
