+++ src/main/java/com/goggles/mentoring_service/domain/category/exception/InactiveCategoryCannotMoveException.java
@@ -1,7 +1,7 @@
 package com.goggles.mentoring_service.domain.category.exception;
 
 
-import com.goggles.mentoring_service.domain.common.MentoringValidationException;
+import com.goggles.mentoring_service.domain._common.MentoringValidationException;
 
 public class InactiveCategoryCannotMoveException extends MentoringValidationException {
 	public InactiveCategoryCannotMoveException() {
